+++ spec/support/serializers.rb
@@ -1,7 +1,0 @@
-class PersonSerializer < ActiveModel::Serializer
-  attributes :name
-end
-
-class ArticleSerializer < ActiveModel::Serializer
-  attributes :title
-end